Add files via upload
Problem Statement:¶
Shoes store inventory management.

Customer like a shoe but shop have no model that fits customer shoes. Shop offered another shoes but customer didn't like it. As a shop manager, they are not happy too. They lost a customer — it’s you. And if this is the first time you visited the store, will you come back? Of course… “maybe”. But if this is the second time you got the same problem with this store, will you come back? It’s absolutely a shop manager’s nightmare.

So how can we increase the chance that customer will find their favorite shoes at our store? what is the number of shoes (model, size) that are likely to be sold?

Sol: Analysis done on 2016 data. To increase the chances of customer getting their favorite shoes.